requirements.txt
absl-py==0.8.1
alembic==1.3.0
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gast==0.2.2
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.24.3
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mtcnn==0.0.9
numpy==1.17.3
oauthlib==3.1.0
opencv-contrib-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.10
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
